Future Learning + Dependency Inlining!
I think Mike Keneally 'Cram' might be might favorite song of his. Was listening to it last night on my old iPod and with my earbuds, the artistic choices are outstanding, and the mix is perfection. I can only strive to make a song like that. Best song off of Inkling.

https://www.youtube.com/watch?v=aKRKJp5m3Gc
https://www.youtube.com/watch?v=yeStc2ZuuOs
https://www.youtube.com/watch?v=nE6UUg0Nenc

After looking and thinking about the artwork for Empath (and now come to think of it, Inkling), I want to experiment with having a foreground color, and background text cutout kind of concept, it looks so cool!

Listening to Cornbread Crumb right now :)