Vertical Grids
Psh Psh, Voyager

Finally figured out that that was where that riff was familiar from! I heard it on Fox first, then on M&M it sounded familiar, but I didn't know where from.

It's not the same riff really, but this made me think of Ignorant Elephant. I think my music kind of has a similar kind of thing going to that one, possibly the feng shui of it (that's a new term I heard the last few days haha).