Symphony X Lineup Reference
Adding this in case the lineup changes down the road, or if the URLs to the page change for some reason (the internet is a crazy place, think of all the hullabaloo around .io domains right now!!).

https://en.wikipedia.org/wiki/Symphony_X#Line-up

Can't believe how accurate the current demo is for this! Now I just have to make it into more of a nice abstraction, then I will make it work for my discography too.