rBnaodn Re-data rendering!
I want to re-add hover tooltips for each album though, as well as add overlapping graphs so I can show when certain instruments or recording techniques were used over time. Only a small subset of projects of mine have vocals for example, and I got my bass guitar in fall 2022, so that changed things up a lot for example. The later part of this year was all about mixing and EQ, next up seems like it might be with mastering, but that's more of foreshadowing than anything else 😊